+++ datanode/broker/convert.go
@@ -129,6 +129,8 @@ func toEvent(ctx context.Context, be *eventspb.BusEvent) events.Event {
 		return events.SnapthostTakenEventFromStream(ctx, be)
 	case eventspb.BusEventType_BUS_EVENT_TYPE_DISTRESSED_ORDERS_CLOSED:
 		return events.DistressedOrdersEventFromStream(ctx, be)
+	case eventspb.BusEventType_BUS_EVENT_TYPE_EXPIRED_ORDERS:
+		return events.ExpiredOrdersEventFromStream(ctx, be)
 	}
 
 	return nil